+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-11-24
+
+### Added
+- add tests for typing and combining translations
+
+### Removed
+- Fix typing to be compatible with strict typescript
+
 ## [0.2.2] - 2025-11-24
 
 ### Changed